Update patch set 2
Patch Set 2:

(15 comments)

This change is conflicting with https://gerrit.onap.org/r/#/c/85210/ - change that fixes now broken PnfEventReadyDmaapClient. The conflicts are complex that it is hard to merge it without rewriting.

This change should be done in new block in bpmn, instead of changing logic in external (from the bpmn point of view) class - PnfEventReadyDmaapClient.

After this commit it's hard to actually understand what's going on in PnfEventReadyDmaapClient and it makes impossible to make any changes without accidentally breaking it.

At first glance it seems that this change - i.e. communication with aai - is something that should be implemented as a separate bpmn task since everything is unrelated and disconnected functionally from the surrounding Dmaap logic. A dedicated bpmn task would make it clear that the workflow is communicating with AAI. Also the violation of the single responsibility principle could be avoided in DmaapClient.

Patch-set: 2
Reviewer: Gerrit User 2795 <2795@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Verified=0