Update patch set 6
Patch Set 6:

Hi Venkat. Thanks for updating the commit message. It was much clearer. As for abandoning a change to fix the build, we normally would continue to work on the same change and update it until it was ready to merge. For committers (I am one) it is very helpful to have the full start to finish of how a change was evolved, comments etc. Also it clogs up gerrit to have multiple changes that cover the same topic, with some abandoned etc. Next time please continue working the same change as there is little reason you would need to create a new one, regardless of commit changes, code changes etc.

Patch-set: 6