Update patch set 1
Patch Set 1:

A few more questions about this since the intent of the current implementation using * is not clear.

As we know the asterisk character is not a wildcard in SQL, but simply a char. the '%' in conjunction with LIKE clause serves that purpose. 
- So having the asterisk replace all NULL values is just an interpretation and marking of that empty value with something "nice"? 

- How are the asterisks values populated in the DB in the first ? Is it something that must be done manually (by doing the suggested UPDATE) before service instantiation ?
Besides this method, we have not found any other places that rely on the asterisk (neither persisted nor interpreted in the code)

You have mentioned that the UPDATE is the least impacting fix - What would it actually impact (or break) if we simply do not do a select with the asterisk character in the WHERE clause?

Basically if the service type column has to be updated manually each time just to meet the asterisk criteria of the select clause, then it is an issue that should be fixed quickly since it breaks the instantiation flow with an ambiguous, hidden requirement (the update). In this case it may be a cleaner and less complicated solution to rely on NULL values and just ignore the servicetype altogether. The result will be the same - all servicetypes will be considered/included when none are given as criteria. 
Another question/issue (for a future improvement) is if there should be a value introduced for representing ALL service types. This would probably require introducing an aggregated service type value, making the column and entity field serviceType not nullable. It may be a good approach since currently the serviceType values have the following meaning: NULL == '*' == ALL

Patch-set: 1
Reviewer: Gerrit User 2815 <2815@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Verified=0