Update patch set 1
Patch Set 1:

Thanks Steve for explanation. Now it is clear what "*" means in this context. I thought that this query is supposed to choose the NorthBoundRequest object regardless the value of serviceType, that "*" means any. But as you said, this query should not choose NorthBoundRequest that has specific value of serviceType, for example "transport". 
So, the "*" in the table means that NorthBoundRequest is for any serviceType. If we put "null" value instead of "*", it would be misleading, because null means that there are no serviceType for this NorthBoundRequest.

In the other hand the current form can be misleading and should be improved later in El Alto release.

Patch-set: 1
Reviewer: Gerrit User 2789 <2789@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0