Update patch set 1
Patch Set 1:

Hey Marios. You don't need to abandon a change to update it. In fact that is highly frowned upon, as we don't have the history of the change all in one place. On the next change please update the change rather than abandoning it and creating a new one. If you need pointers on how to do this with gerrit, google is good place to start. Or you can ask on the mailing list.

Patch-set: 1