Update patch set 1
Patch Set 1:

Hi Steve, the focus of this is authentication rather than authorization. For El Alto we are planning to make use of spring boot support for authentication.
We can of course consider moving to using AAF when we get a better understanding of how to use AAF to perform the authentication that we require (TLS with tokens, two way TLS, TLS with basic auth) and a better understanding of how an AAF based solution will work with an external component (VNFM)

Patch-set: 1
Reviewer: Gerrit User 5269 <5269@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0