Update patch set 1
Patch Set 1:

Hi Steve, here is my two cents... The reason I asked you if the concern is about  authentication or authorization is that authorization support properly is not trivial. It is good that AAF supports the URL-pattern based authorization. To me, it is a coarse-grained authorization. AAF documentation said it supports fine-grained authorization, but I need to know more details.
Also, authorization is enforced by the service provider (e.g., PEP, PDP) based on the roles, attributes or policies (RBAC, ABAC, PBAC). AAF can control the ONAP internal components, but in this case, we are dealing with external vendor SVNFM, so the vendors have their own authorization mechanisms. In short, we need a better understanding of AAF and overall ONAP authentication and authorization strategy.   Maybe, we approve this as the first step and open another discussion for security. Then, we apply the common solution to the whole ONAP SO. Once you share AT&T security design and implementation, we can review together.

Patch-set: 1