Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> (1 comment)
> 
> suggestion to recheck

There is no way this is currently set in an automated way in on the platform.  Eventually it should be set in SDC, but today it is not.  So either we have to manually insert record every time or default it to a sensible value.

Patch-set: 2
Reviewer: Gerrit User 63 <63@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Verified=0