Update patch set 1
Patch Set 1: Code-Review-1

Hi, I would like to explain why this ticket should not be merged:

1. Adding logger message to method.

You put logger messages to methods like this

public void someMethod() {
	logger.debug("someMethod start");
	...
	catch() {
	   ...logger.degub("ex thrown")
	}
	...
	logger.debug("someMethod end")
}

I think that logger with "start" and "end" method are pointless, you do not even print any value that take a part in the flow, it is just a marker that method starts and ends. This makes a mess, debug.log file contains a lot more, but useless information.

In this case, the logger in catch block is sufficient, because then we have an info that some ex occurred. If you add logger to catch clauses, please print exception message also.

2. Adding java doc

Most of java doc that you added is like this:

  /**
     * this method prints values
     *
     * @param value1
     * @param value2
     */
public void printValues(String value1, String value1) {
	...
}

These kinds of java docs are useless, the description does not tell more than method name. There is no added value with that. If you want to add java doc to method, the description should tell a little more, otherwise it is a trash in the code.

Patch-set: 1
Label: Code-Review=-1