Update patch set 1
Patch Set 1:

Why would we need to keep different versions?
If a new version becomes available we either switch to the new version (if there is a reason to), or we stay with the current version.

Patch-set: 1
Reviewer: Gerrit User 5269 <5269@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Verified=0