Update patch set 1
Patch Set 1: -Code-Review

Hi Seshu and Steve - this is an update of the ETSI Catalog Manager swagger. This original one was merged under SO-2501. After then, we found missing APIs and the CMCC/ZTE provided us a new swagger file. 
Could you please review it and approve if it is good?

Patch-set: 1