Update patch set 13
Patch Set 13:

> Patch Set 13: Code-Review-1
> 
> (2 comments)

Can we get clarification about this one? in review https://gerrit.onap.org/r/c/so/+/98166, it's not specified. And in the ppt slides attached in https://jira.onap.org/browse/SO-2312, it's mentioned as 'module'. 

in this code review, scope is fetched from execution context. Even it's 'vf_module' or 'module' in the orchestration table, some transformation can be done before calling the ControllerExecutionBB/DE with expected value 'pnf/vnf/vf' for scope.

Patch-set: 13