Encryption and decryption of OpenStack Passwords removed
Description from Jira ticket:

Current way of passing OpenStack password to SO is to
encrypt it with a symmetric key that can be find in both
OOM repo and source code. This means that the key is already
compromised and this additional encryption does not introduce
any additional security layer. Additionally this creates a fake
security feeling for people who are not very familiar with the
cryptography as they may think that their password is safe because
it has been encrypted.

Instead of using some custom-made method of passing this password
please just use a secret and pass it via environment variable
unencrypted
and just never store it in a config file

Issue-ID: SO-2535
Change-Id: I592e0ed17891180c329d43c521a2e3f60908c480
Signed-off-by: Dominik Mizyn <d.mizyn@samsung.com>