Update patch set 1
Patch Set 1:

@Steve Smokowski
How exposing a password encrypted with a compromised symmetric key is more secure than just exposing it in a plain text?

There has to be a way to pass a secret material from outside to the application container. If you believe that environment variable is not a good idea for SO please propose any other alternative form that is supported by kubernetes is is going to work better in this case

Patch-set: 1
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Verified=0