add retry sequence levels
Added retry sequence levels so that the different services can specify
their desire.
Updated to be medium and added to be long sequence in audit.
Change retry level to be final and changed default constructor to
private
Changed default constructor to public so its visible.
Updated unit test to reflect change and get verify job to pass
Added and updated the unit test for add graph

Issue-ID: SO-2649
Signed-off-by: Benjamin, Max (mb388a) <mb388a@att.com>
Change-Id: I981b90a5dbb1dca5354fba37589b26dabbfc0b0e