Update patch set 3
Patch Set 3:

@Sylvain, @Krzysztof - As you know, we are following the current SO security mechanism. There was existing onap-ca.crt in the docker-files/ca-certificates directory, and we are adding mbs-ca.crt for MSB secure communication. 
As you said, we need a better solution for SO and other ONAP components, and asked the direction to SECCOM. I don't think that happens for Frankfurt. It is too late. I think we need to wait for Guilin and we all need to discuss a proper security solution for Guilin.
In my opinion, we should follow the current SO security mechanism and enhance it in the Guilin after SECCOM direction.

Patch-set: 3