Update patch set 3
Patch Set 3: Code-Review-1

My take on this is clear:

*No hard coding certificates in docker images in any case.* 

It's basically the worst thing that you can do.

I understand that this is the current state in SO but:
1) it doesn't mean that I accept it (I don't but I just don't have enough cycles to fix all bad constructions in ONAP code base at once)
2) It doesn't mean that you have to make things even worse by adding yet another hard coded certificate

As Sylvain said, the prefer way would be to use AAF init container. If you get a waiver from seccom/tsc we can discuss adding the certificate to OOM repo but never ever to the docker image.

Patch-set: 3
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1