Update patch set 3
Patch Set 3:

I am all in favor of pulling the certificates out of the container BUT code was merged to make the running use SO without solving this issue so now all testing is blocked until we either bandaid the certificate issue or revert the change to run the container as a non-root user. Pick one but do it fast.

Patch-set: 3