Update patch set 3
Patch Set 3:

I second Brian Freeman. Beside of this bug, the start-app.sh fix for non-root user broke the certificate. @Sylvain, Seshu said he opened a ticket for this. Could you please fix it ASAP.  @Seshu, please let me know your advice. Since it is blocking SO test, I suggested to fix the security with a better solution under a separate use case although I understand the Krzysztof and Sylvain's point. If our solution is still not acceptable, we need to revert this (by the way this code is not merged, but the start-app.sh for non-root user was merged).
Please let us know.

Patch-set: 3