Update patch set 3
Patch Set 3:

first of all, what we are reverting? This code made by Waqas has not been merged. Only code we want to revert is the start-app.sh which is not part of this bug.
secondly, to my knowledge, MSB certificate is not handled by AAF yet (which is for Guilin I think as I saw the MSB certificate plan for AAF migration).
FYI, this code is not acceptable, most likely, the SOL003 Adapter will not use MSB to communicate with the ETSI Catalog Manager.

To me, the start-app.sh fix is the first to make SO test running.

Patch-set: 3