Update patch set 3
Patch Set 3:

I think Brian confirmed, https://gerrit.onap.org/r/#/c/so/+/102978/ start-app.sh (non root-user) should be fixed first regardless this bug. It broke the existing onap-ca.crt handling. 
Secondly, the MSB issued the self-signed certificate according to Huabing (MSB PTL). We got the MSB certificates from him, and he pointed the user story that MSB has a migration path for AAF in Guilin.  I don't know their story with TSC/SECCOM.
Thirdly, the security solution in ONAP is not yet production-qualify in my opinion. So, we are doing based on what we have. If our code is not acceptable, we need to treat the communication between the SOL003 Adapter and ETSI Catalog as an internal communication. But I hope we want to make the communication secure, so in Guilin, we know what to challenge. Not only this, the SOL003 Adapter and external SVNFM communication could be much complicated - we put this one as the Guilin work item.

Patch-set: 3