Orchestration transitions missing for post instantiation workflow
This is a list of transition states that were added internally at Bell when working with the BB workflow and post instantiation.

Those are to support Rollback and Delete even if the VNF is in the status ConfigAssigned, Configure or Configured.
The missing trasition states was causing the Rollback or Delete to fail.

Also update the initial script to support already existing data.

Issue-ID: SO-2798
Signed-off-by: Sebastien Premont-Tendland <sebastien.premont@bell.ca>
Change-Id: Ic90b74a25a5790f6d9ddb6919771735b1c081cb5