Fix transition states for the BB workflow.
Deleted the versioned flyway file in favor of the repeatable
one for MacroData. I know deleting a versioned file could mess up
upgrade but given no release has been done with those new versioned files
it should be ok and cleaner.

This is a list of transition states that were added internally at Bell
when working with the BB workflow and post instantiation.

Those are to support Rollback and Delete even if the VNF is in the status
ConfigAssigned, Configure or Configured.
The missing transition states was causing the Rollback or Delete to fail.

Issue-ID: SO-2798
Signed-off-by: Sebastien Premont-Tendland <sebastien.premont@bell.ca>
Change-Id: I21c11298a456a472cedc7a8733c6dbc3a071f3a6