Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> This is fine for  the master part , for release would recommend to change to the release version instead of snapshot

I'm sorry to say that but it's not fine not SNAPSHOT on master. 
SNAPSHOT are for test purpose and then people should release and use this on master.
Especially with LF policy of removing SNAPSHOT after a while (which is fine)

This part of "Don't break the build"

Patch-set: 5