Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> (1 comment)
> 
> we should wait for gate to be working again,
> we also should use ONAP Java11 version

The idea here on this review is to "prepare" the code base without requiring the infrastructure to switch to java11. There is a java11 profile which will switch the compile version. We have tested this in AT&T on openjdk 11.0.6 and java8. All of these changes are to make sure that we're using dependencies which support java 11.

Patch-set: 2