Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2: Code-Review-1
> > 
> > (1 comment)
> > 
> > we should wait for gate to be working again,
> > we also should use ONAP Java11 version
> 
> The idea here on this review is to "prepare" the code base without requiring the infrastructure to switch to java11. There is a java11 profile which will switch the compile version. We have tested this in AT&T on openjdk 11.0.6 and java8. All of these changes are to make sure that we're using dependencies which support java 11.

Not that not against the change you propose (I'm actually a lot in favor).

I just saw that just so-simulator image is changed, my bad, I thought it was for the whole images.

Nevertheless, using ONAP base java image from integration on so-simulator and validating this (huge) change with gating seems to be the way to go, at least for me.

Patch-set: 2
Reviewer: Gerrit User 2319 <2319@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>