Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > This change removes the autowired annotation making this class throw null pointers when not under test.
> 
> Hi Max, as of Spring 4.3, classes with a single constructor can omit the @Autowired annotation.

Okay no problem. It's just that all the other classes use @InjectMocks in the unit test for @Autowired fields and this one was a different style. I am not sure if we should have two styles of doing this. I will be putting up a changeset with @InjectMocks and we can discuss it further.

Patch-set: 1
Reviewer: Gerrit User 97 <97@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>