Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> Hello, 
> 
> I believe it's really infortunate not to try to understand why gating failed here.
> You also didn't gave a chance to have a technical discussion on why this test has failed (knowing it never fails usually...) by commenting and merging in the same minute.
> 
> I also understand that gating is a waste of resource here. And as these resources are constrained (we're holding CLAMP to use them for example), I prefer not to waste them here and use them for project with real will to use it.

I am happy to have a discussion about why gating failed but I am not provided enough information to act upon it. Where are the logs from the robot run? Where is the source code for the tests? Can you point me to a wiki for these?

I have no issue troubleshooting your CI for weeks (https://jira.onap.org/browse/SO-2941) but as discussed master is not stable or gating for SO. This is our next release and there may be some issues in the interim. I must have missed the vote when we agreed to allow this CI process to gate SO on master. Gate SO on Frankfurt, Dublin, Casablanca, Berlin, Amsterdam, where there's an expectation of stability.

Patch-set: 9