Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> > Patch Set 9:
> > 
> > Hello, 
> > 
> > I believe it's really infortunate not to try to understand why gating failed here.
> > You also didn't gave a chance to have a technical discussion on why this test has failed (knowing it never fails usually...) by commenting and merging in the same minute.
> > 
> > I also understand that gating is a waste of resource here. And as these resources are constrained (we're holding CLAMP to use them for example), I prefer not to waste them here and use them for project with real will to use it.
> 
> I am happy to have a discussion about why gating failed but I am not provided enough information to act upon it. Where are the logs from the robot run? Where is the source code for the tests? Can you point me to a wiki for these?

On any gating attempt, you get an test results url in gerrit
in your case the link was https://gating-results.onap.eu/results//107820-9/index.html
it is no more available as you merged (logs are cleaned once the patch is merged)
but you can find some examples
the link shall lead you to a page like this one: https://gating-results.onap.eu/results/onap_daily_pod4_frankfurt-605055463-06-22-2020_02-06/
if you click on the different test components, you can have access to all the logs corresponding to the tests
- onap-k8s => https://gating-results.onap.eu/results/onap_daily_pod4_frankfurt-605055463-06-22-2020_02-06/infrastructure-healthcheck/k8s/index.html
here you can even have access at the log of each pod (at the end of the deployment)
- healtchcheck: https://gating-results.onap.eu/results/onap_daily_pod4_frankfurt-605055463-06-22-2020_02-06/healthcheck/full/full/report.html
- healthdist (onboarding/distribution of the vFW) - test that was broken by the change: https://gating-results.onap.eu/results/onap_daily_pod4_frankfurt-605055463-06-22-2020_02-06/healthcheck/healthdist/healthdist/report.html
- end to end onboarding/distribution/instantiation: https://gating-results.onap.eu/results/onap_daily_pod4_frankfurt-605055463-06-22-2020_02-06/smoke-usecases/basic_vm/xtesting.debug.log
....

several presentations have been done since we put this system in place in February 2019 (DDF, PTL meetings, TSC meetings, Integration meetings)
An Integration Q&A is available here: https://wiki.onap.org/display/DW/OOM+Gating
It is also part of the official ONAP Integration documentation: https://onap-doc.readthedocs.io/projects/onap-integration/en/frankfurt/onap-integration-ci.html#integration-ci

The code for the healthcheck is the one you can find in the onap/testsuite repository

> 
> I have no issue troubleshooting your CI for weeks (https://jira.onap.org/browse/SO-2941) but as discussed master is not stable or gating for SO.
This is our next release and there may be some issues in the interim. I must have missed the vote when we agreed to allow this CI process to gate SO on master. Gate SO on Frankfurt, Dublin, Casablanca, Berlin, Amsterdam, where there's an expectation of stability.

As far as I remember, it was decided to add Gating to SO with Seshu during the ONS or the DDF.
Several projects asked to get such feature, due to the fact that we did not have enough resources (full ONAP deployment on each patchset submission), it was decided to priviledge OOM (installation) and SO (critical core component)

As Gating is a tool to give feedback to developers by deploying an OnDemand ONAP per patch. It rebuilds the docker then deploy and test. You have access to the logs of lots of tests dealing with infrastructure, component healthcheck, end to end tests and security tests. 
Gating Master is a usual process in open source projets. Master does not mean necessarily unstable. It is a dev branch, feature can be incomplete but there is no reason in this case for instance that the change prevents the vFW to be distributed..it maybe reports a regression or an interesting side effect that is always better to fix before any merge.

If you look at the last daily Master over the last 7 days, it is pretty stable (look at the end https://gating-results.onap.eu/results/)

Patch-set: 9
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>