Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-2
> > 
> > (1 comment)
> 
> I'm not saying it's right or wrong, this is just the style we have through out the codebase. We can start to change it, but I would rather start that as a separate effort.
> 
> Here are some samples of the pattern already in use:
> https://github.com/onap/so/blob/40396c8763a600059d025fb202802ac70d406b08/adapters/mso-openstack-adapters/src/test/java/org/onap/so/adapters/tenant/MsoTenantAdapterImplTest.java#L41
> 
> https://github.com/onap/so/blob/e141aad67e997015bbb6798c790e6702d1894115/bpmn/mso-infrastructure-bpmn/src/test/java/org/onap/so/bpmn/common/workflow/service/WorkflowAsyncResourceTest.java#L22
> 
> https://github.com/onap/so/blob/32464a960c6d26fa6a3a1bb25475ac2c1b96e829/common/src/test/java/org/onap/so/utils/ExternalTaskServiceUtilsTest.java#L26
> 
> https://github.com/onap/so/blob/01d814c45616bcde613db3ca760df5c9348337a5/mso-api-handlers/mso-api-handler-common/src/test/java/org/onap/so/apihandler/common/ResponseHandlerTest.java#L41
> 
> https://github.com/onap/so/blob/515be95fb12dcecf7cffb9f8fac6db853907d329/mso-api-handlers/mso-api-handler-infra/src/test/java/org/onap/so/apihandlerinfra/infra/rest/validator/NetworkDeleteValidatorTest.java#L25

Hi Max,
I agree that we should follow one style in tests, using Mockito annotations like @InjectMocks. No problem with that. 
But my comment is for @Spy anno in this particular test class, which means we declare tested object as real object and mock in the same time.
In these tests I do not need to declare spy object, this anno is useless and misleading for code readers.

Generally I think that @Spy should be used carefully, only in particular cases where it is really needed.

I saw that in some tests there is @Spy anno although it is useless, and it is wrong. Do not put anno that you are not using.

All I want is to remove this one anno - @Spy, from this class. All tests pass because tested object does act as mock in any of these tests. The rest of changes that you made are ok.

Patch-set: 1
Reviewer: Gerrit User 2789 <2789@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>