Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > (7 comments)
> > 
> > > Patch Set 3:
> > > 
> > > Why did we lose the fluent builder feature of ExecuteBuilding block in this change?
> > 
> > I removed this, because setters chaining can get problematic when used with inheritance.
> > You can read about it here: 
> > https://stackoverflow.com/questions/14105275/fluent-setters-with-inheritance-in-java
> > I came up with inheritance because I want to remove code duplication and this is still work in progress.
> 
> Okay, we can add it back. I can write up a demonstration that solves the inheritance issue. This was already accepted though. It will be in another review.

I think that if the object is not immutable or is not created via contructor with many arguments (builder pattern), there is not added value with fluent setters, it is only a matter of code style.

Of course, as we spoke before in different gerrit ticket, community should have common way of code style, and we agreed to that.
However, I do not see in SO code that community follows the fluent setters pattern. 
I accepted this ticket because for me, if there is no reason for fluent setters, the void setters are more clean to me. Question to you Max, why it is important to you to get fluent setters back?

Patch-set: 3