Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > (7 comments)
> > > 
> > > > Patch Set 3:
> > > > 
> > > > Why did we lose the fluent builder feature of ExecuteBuilding block in this change?
> > > 
> > > I removed this, because setters chaining can get problematic when used with inheritance.
> > > You can read about it here: 
> > > https://stackoverflow.com/questions/14105275/fluent-setters-with-inheritance-in-java
> > > I came up with inheritance because I want to remove code duplication and this is still work in progress.
> > 
> > Okay, we can add it back. I can write up a demonstration that solves the inheritance issue. This was already accepted though. It will be in another review.
> 
> I think that if the object is not immutable or is not created via contructor with many arguments (builder pattern), there is not added value with fluent setters, it is only a matter of code style.
> 
> Of course, as we spoke before in different gerrit ticket, community should have common way of code style, and we agreed to that.
> However, I do not see in SO code that community follows the fluent setters pattern. 
> I accepted this ticket because for me, if there is no reason for fluent setters, the void setters are more clean to me. Question to you Max, why it is important to you to get fluent setters back?

I can go one way or the other. I think it's more about the conversation about scope and intention. I think creating a base class is a good change, but then it led to non-fluent setters because "it can get problematic". This is an interesting reason. It is not necessarily true that Mateusz wanted to remove fluent setters but perhaps felt like it was the only path to getting a common base class.

I think it's a good opportunity to talk about how to maintain fluent setters with inheritance.

The history of this change and why we have fluent setters goes back a few years. We were looking to eventually move towards a fluent builder pattern, but a lot of the code requires you to mutate a calling object and we didn't have time to change everything. This was left as to come back later and further convert it. It is nice with these objects that have a lot of sets to compactly chain them and have your IDE autocomplete them.

If this was part was part of an effort to "remove fluent builders from the so codebase" then we'd have a different discussion. I feel like this was well meaning side effect of another refactor and I'm not sure I agree with it. Also, the review was still WIP so I didn't look at it yet 😄

Patch-set: 3