Update patch set 4
Patch Set 4: Code-Review-1

What these all not null assertions are supposed to check? In each case you check the if the object or value is null and all this objects and values are set at the test beginning. These object are not even test results but prerequisites for tests that are set.
For me these assertions are useless.

Patch-set: 4
Reviewer: Gerrit User 2789 <2789@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1