Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > How come it was an issue that ConfigBuildingBlocksDataObject needed to implement serializable?
> 
> Well, it is because object that does not need to implement serializable interface - shouldn't be forced to do so. This is just bad design and can be confusing for others.

I think the situation was that the base was not serializable when it needed to be. We kept the refactor internally and it eventually broke our integration tests. I think it's okay to force serializable on these pojos. Most things in the BPMN space should be serializable. There are a number of additional commits coming as there have been a bunch of issues found in the WorkflowAction area. We'll revisit dropping this after all the downstream commits fixing other stuff goes in.

Patch-set: 4
Reviewer: Gerrit User 97 <97@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>