Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Also no need to do update just change the initial insert. The R__MacroData gets executed every time and at the top it deletes everything from the table.

Yes and no. If you will look into thr history, the script maintains the history of changes also. If i will change the initial insert it might be upated alter on. I don't tsay this is the case now but this is the rule in this script, thats why changes are added on the bottom to be sure they are not covered by further updates  later on in the script.

Patch-set: 2