Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Just a question... Is there any reason it used Groovy for all the business logic, instead of Java logic? In SO, we try to get rid of Groovy dependencies.

Hi Byung-Woo Jun,
In SO, we have a groovy script corresponding to every bpmn process since its inception. Not sure of the reason but if we want to completely move to java, I suppose there will be one java delegate per block in the bpmn process which might be difficult to maintain.

Patch-set: 4