Update patch set 5
Patch Set 5: Code-Review-2

This is not the solution of this problem.

The problem is that we get the list of orchestation flows sorted by primary id, not by sequenceNumber.
SequenceNumber defines the order of building block execution.

The problem appears after the following commit:
https://gerrit.onap.org/r/c/so/+/111019

After inserting the ControllerExecutionBB into table, it is on the last record of the table orchestation_flow_reference. 

And that is why when we get the list of bb in WorkflowAction class, ControllerExecutionBB is the last,
but it can not be the last, it must be executed after waitForPnfBB, because this is defined by sequence number.

Here is a solution corrected by me, it is tested on lab and works.
https://gerrit.onap.org/r/c/so/+/114127

Patch-set: 5
Reviewer: Gerrit User 2789 <2789@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-2