Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review-2
> 
> This is not the solution of this problem.
> 
> The problem is that we get the list of orchestation flows sorted by primary id, not by sequenceNumber.
> SequenceNumber defines the order of building block execution.
> 
> The problem appears after the following commit:
> https://gerrit.onap.org/r/c/so/+/111019
> 
> After inserting the ControllerExecutionBB into table, it is on the last record of the table orchestation_flow_reference. 
> 
> And that is why when we get the list of bb in WorkflowAction class, ControllerExecutionBB is the last,
> but it can not be the last, it must be executed after waitForPnfBB, because this is defined by sequence number.
> 
> Here is a solution corrected by me, it is tested on lab and works.
> https://gerrit.onap.org/r/c/so/+/114127

I think both solutions are needed as we need to have a check in place to make sure we don't go out of bounds even if we assume ControllerExecutionBB will not be the last BB. Also the Listener should be called in a separate task for clarity as it updating the flow. Also this change is needed as, future flow manipulators might not have the same assumption that SkipCDSBuildingBlockListener has.

Patch-set: 5